loyalty point creation for client float field

+++ Client/models.py
@@ -378,7 +378,7 @@ class LoyaltyPoints(models.Model):
     loyaltytype = models.CharField(choices=LOYALTY_CHOICE, default='Service' , verbose_name='Loyalty Type', max_length=50)
     amount_spend = models.PositiveIntegerField(default=0, null=True, blank=True)
     number_points = models.PositiveIntegerField(default=0, null=True, blank=True)
-    earn_points = models.PositiveIntegerField(default=0, null=True, blank=True)
+    earn_points = models.FloatField(default=0, null=True, blank=True)
     total_earn_from_points = models.PositiveIntegerField(default=0, null=True, blank=True)
     
     is_active = models.BooleanField(default=True)
